+++ nbproject/project.properties
@@ -75,7 +75,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=QLTTTH
-main.class=qlttth.QLTTTH
+main.class=qlttth.Class.QLTTTH
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
