+++ src/qlttth/Panel/panelAddNewClass.form
