+++ src/qlttth/Panel/panelManageCourse.java
@@ -2,18 +2,18 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JPanel.java to edit this template
  */
-package qlttth;
+package qlttth.Panel;
 
 /**
  *
  * @author daoho
  */
-public class ManageCourse extends javax.swing.JPanel {
+public class panelManageCourse extends javax.swing.JPanel {
 
     /**
      * Creates new form ManageCourse
      */
-    public ManageCourse() {
+    public panelManageCourse() {
         initComponents();
     }
